# Searcher: reduce allocations for chunk matches