--- a/cmd/searcher/internal/search/search_structural.go
+++ b/cmd/searcher/internal/search/search_structural.go
@@ -148,7 +148,7 @@ func chunkRanges(ranges []protocol.Range, interChunkLines int) []rangeChunk {
 			// First iteration, there are no chunks, so create a new one
 			chunks = append(chunks, rangeChunk{
 				cover:  rr,
-				ranges: []protocol.Range{rr},
+				ranges: ranges[:1],
 			})
 			continue
 		}
@@ -156,7 +156,7 @@ func chunkRanges(ranges []protocol.Range, interChunkLines int) []rangeChunk {
 		lastChunk := &chunks[len(chunks)-1] // pointer for mutability
 		if int(lastChunk.cover.End.Line)+interChunkLines >= int(rr.Start.Line) {
 			// The current range overlaps with the current chunk, so merge them
-			lastChunk.ranges = append(lastChunk.ranges, rr)
+			lastChunk.ranges = ranges[i-len(lastChunk.ranges) : i+1]
 
 			// Expand the chunk coverRange if needed
 			if rr.End.Offset > lastChunk.cover.End.Offset {
@@ -166,7 +166,7 @@ func chunkRanges(ranges []protocol.Range, interChunkLines int) []rangeChunk {
 			// No overlap, so create a new chunk
 			chunks = append(chunks, rangeChunk{
 				cover:  rr,
-				ranges: []protocol.Range{rr},
+				ranges: ranges[i : i+1],
 			})
 		}
 	}